+++ src/main/java/fri/shapesge/engine/GameOnCloseOperationType.java
@@ -1,0 +1,8 @@
+package fri.shapesge.engine;
+
+public enum GameOnCloseOperationType {
+    HIDE,
+    EXIT,
+    DO_NOTHING,
+    SEND_MESSAGE
+}